Resolved Constructor issue
The AddressBookManager had not properly provided with constructor parameter which was conflicting with connection object ,so we made it as static connection = null.
Changed the PRIMARY KEY from Auto-incremented ID to Phone Number, because we use Phone number as a key to update and delete contacts.
Issue- The update function does not update the values on the database as well has no window to update the contact
TODO - Resolve the Update function.